split sphere tests on #78
